change getEnemyByType to helper folder
