+++ package.json
@@ -10,6 +10,7 @@
     "eslint-plugin-import": "2.20.x",
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "^5.3.1",
+    "jest": "^26.6.3",
     "stylelint": "13.3.x",
     "stylelint-config-standard": "20.0.x",
     "stylelint-csstree-validator": "^1.9.0",
