+++ src/helper/enemyHelper.js
@@ -1,0 +1,12 @@
+const getEnemiesByType = (scene, type) => {
+  const arr = [];
+  for (let i = 0; i < this.enemies.getChildren().length; i += 1) {
+    const enemy = scene.enemies.getChildren()[i];
+    if (enemy.getData('type') === type) {
+      arr.push(enemy);
+    }
+  }
+  return arr;
+}
+
+export default getEnemiesByType;
