+++ src/helper/laserUpdate.js
