+++ src/scenes/playerInputScene.js
@@ -2,6 +2,7 @@ import Phaser from 'phaser';
 import laderBoardModule from '../APIs/laderBoard';
 import Button from '../objects/button';
 import config from '../config/config';
+import nameForm from '../helper/Form';
 
 class PlayerInputScene extends Phaser.Scene {
   constructor() {
@@ -22,29 +23,11 @@ class PlayerInputScene extends Phaser.Scene {
 
     const gameScore = localStorage.getItem('score') || 0;
 
-    const div = document.createElement('div');
-    const textInput = document.createElement('input');
-    const submitButton = document.createElement('button');
-    textInput.setAttribute('type', 'text');
-    textInput.setAttribute('placeholder', 'Insert you name...');
-    textInput.classList.add('text-input');
-    textInput.required = true;
-    submitButton.textContent = 'Submit';
-    submitButton.classList.add('submitButton');
+    const form = nameForm(this, 'GameOver', gameScore);
 
-    submitButton.addEventListener('click', () => {
-      const name = textInput.value;
-      laderBoardModule.setPlayer(name, gameScore);
-      if (name !== '') {
-        this.scene.start('GameOver');
-      }
-    });
+    document.body.appendChild(form);
 
-    div.appendChild(textInput);
-    div.appendChild(submitButton);
-    document.body.appendChild(div);
-
-    const element = this.add.dom(this.game.config.width * 0.5, -400, div);
+    const element = this.add.dom(this.game.config.width * 0.5, -400, form);
     element.setDepth(100);
 
     const image = this.add.image(300, 400, 'bgImage');
