+++ src/scenes/preloaderScene.js
@@ -25,6 +25,8 @@ class PreloaderScene extends Phaser.Scene {
   }
 
   preload() {
+    window.localStorage.clear();
+
     const progressBar = this.add.graphics();
     const progressBox = this.add.graphics();
     progressBox.fillStyle(0x222222, 0.8);
