+++ src/scenes/titleScene.js
@@ -28,7 +28,6 @@ class TitleScene extends Phaser.Scene {
     this.gameButton = new Button(this, config.width / 2, config.height / 2 + 100, 'spaceButton', 'Credits', 'Credits');
     this.gameButton = new Button(this, config.width / 2, config.height / 2 + 200, 'spaceButton', 'Scores', 'Scores');
 
-    this.model = this.sys.game.globals.model;
     if (this.model.musicOn === true && this.model.bgMusicPlaying === false) {
       this.bgMusic = this.sound.add('menuTheme', { volume: 0.5, loop: true });
       this.bgMusic.play();
